+++ pkg/test/e2eutil/wait_util.go
@@ -55,7 +55,7 @@ func waitForDeployment(t *testing.T, kubeclient kubernetes.Interface, namespace,
 		deployment, err := kubeclient.AppsV1().Deployments(namespace).Get(name, metav1.GetOptions{})
 		if err != nil {
 			if apierrors.IsNotFound(err) {
-				t.Logf("Waiting for availability of %s deployment\n", name)
+				t.Logf("Waiting for availability of Deployment: %s in Namespace: %s \n", name, namespace)
 				return false, nil
 			}
 			return false, err